+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- Added new action and keyboard shortcut to toggle the state of the task on the current line in the Markdown editor. Default hotkey `Ctrl-Enter`
 - Fixed issue with tasks not being collected when the task content starts with a multibyte character
 
 ## 0.5.1